requirements_clean.txt
colorama==0.4.6
cryptography==45.0.5
fido2==2.0.0
Flask==3.1.1
flask_cors==6.0.1
matplotlib==3.10.3
numpy==2.3.1
openai_whisper==20250625
pandas==2.3.1
psutil==6.1.1
PySide6==6.9.1
python_nmap==0.7.1
Requests==2.32.4
scapy==2.6.1
seaborn==0.13.2
setuptools==80.9.0
speechrecognition==3.14.3
winrt==1.0.21033.1
